Crate interbtc_primitives

Modules

issue

oracle

redeem

refund

replace

Structs

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H256Le

Represents a bitcoin 32 bytes hash digest encoded in little-endian

UncheckedExtrinsic

Simple blob to hold an extrinsic without committing to its format and ensure it is serialized correctly.

VaultCurrencyPair

VaultId

Enums

CurrencyId

TokenSymbol

Constants

DOT

INTERBTC

INTR

KBTC

KINT

KSM

Traits

BalanceToFixedPoint

CurrencyInfo

TruncateFixedPointToInt

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Amount

Signed version of Balance

Balance

Balance of an account.

Block

Opaque block type.

BlockId

Opaque block identifier type.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

Index

Index of a transaction in the chain.

Moment

An instant or duration in time.

Nonce

Index of a transaction in the chain. 32-bit should be plenty.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedFixedPoint

The signed fixed point type.

SignedInner

The Inner type of the SignedFixedPoint.

UnsignedFixedPoint

The unsigned fixed point type.

UnsignedInner

The Inner type of the UnsignedFixedPoint.