Crate interbtc_primitives

Modules

issue

oracle

redeem

replace

Structs

BalanceWrapper

a wrapper around a balance, used in RPC to workaround a bug where using u128 in runtime-apis fails. See https://github.com/paritytech/substrate/issues/4641

CustomMetadata

H256Le

Represents a bitcoin 32 bytes hash digest encoded in little-endian

VaultCurrencyPair

VaultId

Enums

CurrencyId

LpToken

TokenSymbol

Constants

BITCOIN_MAINNET

BITCOIN_REGTEST

BITCOIN_TESTNET

DOT

IBTC

INTR

KBTC

KINT

KSM

SECONDS_PER_YEAR

Loans pallet types

Traits

BalanceToFixedPoint

CurrencyInfo

TruncateFixedPointToInt

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

Block

Opaque block type.

BlockNumber

An index to a block.

ForeignAssetId

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

Index

Index of a transaction in the chain.

LendTokenId

Liquidity

Loans pallet types

Moment

An instant or duration in time.

Nonce

Index of a transaction in the chain. 32-bit should be plenty.

Price

Loans pallet types

PriceDetail

Loans pallet types

Rate

Loans pallet types

Ratio

Loans pallet types

Shortfall

Loans pallet types

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBalance

Signed version of Balance

SignedFixedPoint

The signed fixed point type.

SignedInner

The Inner type of the SignedFixedPoint.

StablePoolId

Timestamp

Loans pallet types

UnsignedFixedPoint

The unsigned fixed point type.

UnsignedInner

The Inner type of the UnsignedFixedPoint.